:truck: Renamed instance_creation_failed.py to exceptions.py
