+++ source/YugiAcademy/YuGiOh/booster_packs/tests/test_booster_pack_card.py
@@ -2,7 +2,7 @@ import pytest
 
 from datetime import datetime
 
-from assertions.instance_creation_failed import InstanceCreationFailed
+from assertions import InstanceCreationFailed
 from YuGiOh.booster_packs import BoosterPack, BoosterPackCard
 from YuGiOh.cards import SpellCard, TrapCard, MonsterCard
