+++ source/YugiAcademy/assertions/__init__.py
@@ -1,1 +1,2 @@
 from .assertions import enforce_not_blank, enforce_must_be_between, enforce_must_be_positive
+from .exceptions import InstanceCreationFailed
