+++ source/YugiAcademy/assertions/tests/test_instance_creation_failed.py
@@ -1,6 +1,6 @@
 import pytest
 
-from assertions.instance_creation_failed import InstanceCreationFailed
+from assertions import InstanceCreationFailed
 
 
 def test_exception_raises():
